almost done regular methods in trash

+++ SmartRm/Trash.py
@@ -9,6 +9,7 @@ from datetime import datetime
 import logging
 import Regular
 import re
+import ExeptionListener
 
 # new procedure if trash does not exist ( create new trash)
 
@@ -27,7 +28,7 @@ class Trash(object):
         self.max_capacity = capacity  # the max quantity of files in trash
         self.max_time = time
 
-    def delete_automatically(self, dry_run):  # works
+    def delete_automatically(self, dry_run, verbose):  # works
         # delete the whole trash
         logging.info("Clean the whole trash".format())
         if dry_run:
@@ -42,28 +43,57 @@ class Trash(object):
                         dict_contains = True
                         break
 
-                if dict_contains:
-                    if os.path.isdir(subpath):
-                        shutil.rmtree(subpath)
-                    elif not os.path.isdir(subpath):
-                        os.remove(subpath)
+                try:
+                    if dict_contains:
+                        if os.path.isdir(subpath):
+                            shutil.rmtree(subpath)
+                        elif not os.path.isdir(subpath):
+                            os.remove(subpath)
+                    if verbose:
+                        print 'trash cleaned'
+                except ExeptionListener.TrashError as ex:
+                    logging.error(ex)
+
             logging.info("Clean information about files".format())
             clean_json = open(self.log_writer.file_dict_path, 'w')
             clean_json.close()
             clean_txt = open(self.log_writer.file_dict_path_txt, 'w')
             clean_txt.close()
 
-    def delete_manually(self, path):  # not checked
+    def delete_manually(self, path, dry_run, verbose):  # not checked
         # delete one file manually
+        files_id = self.search_for_all_files_with_this_name(path)
+        if len(files_id) > 1:
+            logging.warning('Found more than 1 file with this name')
+        elif len(files_id) <= 0:
+            logging.warning('There is no such file or directory')
+            # обработать это (когда их больще 1 и меньше 1)
+
         file_id = self.log_writer.get_id(path)
         clean_path = self.get_path_by_id(file_id, self.path)
+        # watch if thete are more? than 1 file with this name
 
-        if os.path.isdir(clean_path):
-            logging.info("Remove directory".format())
-            shutil.rmtree(clean_path)
-        elif not os.path.isdir(clean_path):
-            logging.info("Remove file".format())
-            os.remove(clean_path)
+        # !!!!!!!!!!!!!!!!!1
+        try:
+            if os.path.isdir(clean_path):
+                logging.info("Remove directory".format())
+                if dry_run:
+                    print 'remove directory'
+                else:
+                    shutil.rmtree(clean_path)
+            elif not os.path.isdir(clean_path):
+                logging.info("Remove file".format())
+                if dry_run:
+                    print 'remove directory'
+                else:
+                    os.remove(clean_path)
+                if verbose:
+                    print 'item removed'
+            self.log_writer.delete_elem_by_id(file_id)
+            self.log_writer.write_to_json()
+            self.log_writer.write_to_txt()
+        except ExeptionListener.TrashError as ex:
+            logging.error(ex)
 
     def get_path_by_id(self, file_id, path):  # not checked
         d = os.listdir(path)
@@ -83,7 +113,7 @@ class Trash(object):
                 txt_file = open(self.log_writer.file_dict_path_txt, 'r')
                 print(txt_file.read())
 
-    def restore_trash_automatically(self, dry_run):  # not tested
+    def restore_trash_automatically(self, dry_run, verbose):  # not tested
         # restore the the whole trash
         logging.info("Restore the whole trash".format())
 
@@ -98,23 +128,27 @@ class Trash(object):
                     if _dict['id'] == item:
                         dict_contains = True
                         break
-
-                if dict_contains:
-                    subpath = os.path.split(subpath)  # ???
-                    subpath = self.get_path_by_id(subpath[1], subpath[0])  # ???
-                    logging.info("Restore file".format())
-                    self.restore_trash_manually(subpath, dry_run)
-                    # if os.path.isdir(subpath):
-                    #     shutil.rmtree(subpath)
-                    # elif not os.path.isdir(subpath):
-                    #     os.remove(subpath)
+                try:
+                    if dict_contains:
+                        subpath = os.path.split(subpath)  # ???
+                        subpath = self.get_path_by_id(subpath[1], subpath[0])  # ???
+                        logging.info("Restore item".format())
+                        self.restore_trash_manually(subpath, dry_run)
+                        # if os.path.isdir(subpath):
+                        #     shutil.rmtree(subpath)
+                        # elif not os.path.isdir(subpath):
+                        #     os.remove(subpath)
+                        if verbose:
+                            print 'item restored'
+                except ExeptionListener.TrashError as ex:
+                    logging.error(ex)
             # with
             clean_json = open(self.log_writer.file_dict_path, 'w')
             clean_json.close()
             clean_txt = open(self.log_writer.file_dict_path_txt, 'w')
             clean_txt.close()
 
-    def restore_trash_manually(self, path, dry_run):  # works
+    def restore_trash_manually(self, path, dry_run, verbose):  # works
         # restore one file in the trash
         # check if the path already exists
 
@@ -154,15 +188,19 @@ class Trash(object):
             self.log_writer.delete_elem_by_id(file_id)
             self.log_writer.write_to_json()
             self.log_writer.write_to_txt()
+            if verbose:
+                print 'item restored'
         return None
 
     def check_policy(self, path, dry_run):  # not checked(redo to check the whole bucket)
         logging.info("Check policies".format())
         if self.policy_time:
+            logging.info('time policy')
             confirm = self.check_date_if_overflow(path)
             if confirm:
                 self.delete_manually(path)
         if self.policy_size:
+            logging.info('size policy')
             self.count_size(dry_run)
             pass
 
@@ -256,8 +294,96 @@ class Trash(object):
 
         return suitable_names, suitable_id
 
-    def restore_by_regular(self):
-        pass
+    def restore_by_regular(self, regex, dry_run, interactive, verbose):
+        names, ids = self.get_names_by_regular(regex)
 
-    def clean_by_regular(self):
-        pass
+        for file_id in ids:
+            clean_path = self.get_path_by_id(file_id, self.path)
+            destination_path = self.log_writer.get_path(file_id)
+            new_name = self.log_writer.get_name(file_id)
+            logging.info("Operations with file {file}".format(file=new_name))
+            index = 0
+            for i in reversed(range(len(clean_path))):
+                if clean_path[i] == '/':
+                    index = i
+                    break
+            dirname = clean_path[:(index + 1)] + new_name
+            logging.info("Rename {file}".format(file=new_name))
+            logging.info("Move to original directory {file}".format(file=new_name))
+
+            if os.path.exists(destination_path):
+                logging.warning('Item with this name already exists.id will be added to real name')
+                dirname += file_id
+            if dry_run:
+                print 'rename file and move to original directory'
+                print 'clean record from json'
+            else:
+                if interactive:
+                    ans = ask_for_confirmation(new_name)
+                    if ans:
+                        try:
+                            os.rename(clean_path, dirname)
+                            shutil.move(dirname, destination_path)
+                            self.log_writer.delete_elem_by_id(file_id)
+                            self.log_writer.write_to_json()
+                            self.log_writer.write_to_txt()
+                            if verbose:
+                                print 'item restored'
+                        except ExeptionListener.TrashError as ex:
+                            logging.error(ex)
+                else:
+                    try:
+                        os.rename(clean_path, dirname)
+                        shutil.move(dirname, destination_path)
+                        self.log_writer.delete_elem_by_id(file_id)
+                        self.log_writer.write_to_json()
+                        self.log_writer.write_to_txt()
+                        if verbose:
+                            print 'item restored'
+                    except ExeptionListener.TrashError as ex:
+                        logging.error(ex)
+
+    def clean_by_regular(self, regex, dry_run, verbose):
+        names, ids = self.get_names_by_regular(regex)
+
+        for file_id in ids:
+            clean_path = self.get_path_by_id(file_id, self.path)
+            name = self.log_writer.get_name(file_id)
+
+            if os.path.isdir(clean_path):
+                logging.info("Remove directory {item}".format(item=name))
+                if not dry_run:
+                    if verbose:
+                        print 'remove item'
+                    shutil.rmtree(clean_path)
+                    self.log_writer.delete_elem_by_id(file_id)
+                    self.log_writer.write_to_json()
+                    self.log_writer.write_to_txt()
+                else:
+                    print 'remove item'
+            elif not os.path.isdir(clean_path):
+                logging.info("Remove file {item}".format(item=name))
+                if not dry_run:
+                    if verbose:
+                        print 'remove item'
+                    os.remove(clean_path)
+                    self.log_writer.delete_elem_by_id(file_id)
+                    self.log_writer.write_to_json()
+                    self.log_writer.write_to_txt()
+                else:
+                    print 'remove item'
+
+
+def ask_for_confirmation(filename, silent=False):
+    answer = raw_input('Operation with {filename}. Are you sure? [y/n]\n'.format(filename=filename))
+    if answer == 'n' or answer == 'N':
+        if not silent:
+            print('Operation canceled')
+        return False
+    elif answer == 'y' or answer == 'Y':
+        if not silent:
+            print('Operation continued')
+        return True
+        # sys.exit(self.exit_codes['success'])
+    elif answer != 'y' and answer != 'n' and answer != 'N' and answer != 'Y':
+        ask_for_confirmation(filename)
